+++ back-riae/Models/glosa_fr1.cs
@@ -7,11 +7,11 @@ namespace back_riae.Models
     {
         public int id_glosa { get; set; }
 
-        public int n_recibo { get; set; }
+        public float n_recibo { get; set; }
 
         public int n_voucher { get; set; }
 
-        public int monto_voucher { get; set; }
+        public float monto_voucher { get; set; }
 
         public string nombre_banco { get; set; }
     }
