+++ back-riae/Models/glosa_fr4.cs
@@ -3,7 +3,7 @@
     public class glosa_fr4
     {
         public int id_glosa { get; set; }
-        public int n_recibo { get; set; }
+        public float n_recibo { get; set; }
         public string texto_glosa { get; set; }
     }
 }
